+++ __utils__/mocks/redis/index.ts
@@ -1,0 +1,8 @@
+import Redis from "ioredis-mock";
+
+jest.mock("@lib/integration/redisConnector", () => {
+  const redis = new Redis();
+  return {
+    getRedisInstance: jest.fn(async () => redis),
+  };
+});