+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/edit/components/elements/element-dialog/tests/ElementDialog.test.js
@@ -1,3 +1,6 @@
+/**
+ * @jest-environment jsdom
+ */
 import React from "react";
 import { cleanup, render, fireEvent } from "@testing-library/react";
 import { ElementDialog } from "../ElementDialog";