+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/responses/[statusFilter]/components/Dialogs/line-item-entries/tests/LineItem.test.js
@@ -1,3 +1,6 @@
+/**
+ * @jest-environment jsdom
+ */
 import React from "react";
 import { render, cleanup } from "@testing-library/react";
 import { LineItem } from "../LineItem";
@@ -24,6 +27,9 @@ describe("LineItem", () => {
     await userEvent.click(removeButtons[0]);
     expect(mockCallback).toHaveBeenCalledTimes(1);
     expect(mockCallback).toHaveBeenCalledWith("31-12-XCRV");
-    expect(removeButtons[0]).toHaveAttribute("aria-label", "downloadResponsesModals.lineItemEntries.remove 31-12-XCRV");
+    expect(removeButtons[0]).toHaveAttribute(
+      "aria-label",
+      "downloadResponsesModals.lineItemEntries.remove 31-12-XCRV"
+    );
   });
 });