+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/responses/[statusFilter]/components/tests/DownloadTable.test.tsx
@@ -1,3 +1,6 @@
+/**
+ * @jest-environment jsdom
+ */
 import React from "react";
 import { act, render } from "@testing-library/react";
 import { DownloadTable } from "../DownloadTable";