+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/settings/components/SettingsNavigation.tsx
@@ -1,9 +1,9 @@
 "use client";
-import React from "react";
+
 import { useTranslation } from "@i18n/client";
 import { SubNavLink } from "@clientComponents/globals/SubNavLink";
 import { EmailIcon, BrandIcon, GearIcon, ProtectedIcon } from "@serverComponents/icons";
-import { useFlag } from "@lib/hooks/useFlag";
+import { useIsProduction } from "@lib/hooks/useIsProduction";
 
 export const SettingsNavigation = ({ id }: { id: string }) => {
   const {
@@ -11,7 +11,7 @@ export const SettingsNavigation = ({ id }: { id: string }) => {
     i18n: { language },
   } = useTranslation("form-builder");
 
-  const isAPIEnalbed = useFlag("zitadelAuth");
+  const { loading, isProduction } = useIsProduction();
 
   return (
     <div className="relative flex">
@@ -35,7 +35,7 @@ export const SettingsNavigation = ({ id }: { id: string }) => {
               {t("settings.formManagement")}
             </span>
           </SubNavLink>
-          {isAPIEnalbed.status === true && (
+          {!loading && !isProduction && (
             <SubNavLink href={`/${language}/form-builder/${id}/settings/api`}>
               <span className="text-sm laptop:text-base">
                 <ProtectedIcon className="mr-2 inline-block laptop:mt-[-2px]" />