+++ app/(gcforms)/[locale]/(form filler)/id/[...props]/lib/rehydrateFormResponses.ts
@@ -74,6 +74,6 @@ function _rehydrateDynamicRowResponses(responses: [string, Response][]) {
 }
 
 function _rehydrateCheckBoxResponse(response: Response) {
-const responseParsed = safeJSONParse<{ value: string[] }>(response as string);
+  const responseParsed = safeJSONParse<{ value: string[] }>(response as string);
   return responseParsed?.value || [];
 }