+++ app/(gcforms)/[locale]/layout.tsx
@@ -1,13 +1,21 @@
 import { authCheckAndThrow } from "@lib/actions";
 import { ClientContexts } from "@clientComponents/globals/ClientContexts";
 import { ReactHydrationCheck } from "@clientComponents/globals";
+import { getSomeFlags } from "@lib/cache/flags";
+import { FeatureFlags } from "@lib/cache/types";
 
 export default async function Layout({ children }: { children: React.ReactNode }) {
   const { session } = await authCheckAndThrow().catch(() => ({ session: null }));
+  const featureFlags = await getSomeFlags([
+    FeatureFlags.addressComplete,
+    FeatureFlags.repeatingSets,
+  ]);
   return (
     <>
       <ReactHydrationCheck />
-      <ClientContexts session={session}>{children}</ClientContexts>
+      <ClientContexts session={session} featureFlags={featureFlags}>
+        {children}
+      </ClientContexts>
     </>
   );
 }