+++ lib/tests/auth/cognito.test.ts
@@ -1,6 +1,3 @@
-/**
- * @jest-environment node
- */
 /* eslint-disable @typescript-eslint/no-explicit-any */
 
 import { mockClient } from "aws-sdk-client-mock";